+++ test/seq/StateMapTests/runtests.jl
@@ -3,23 +3,23 @@ module StateMapTests
 using Test
 
 @time @testset "StateMapTests" begin
-  include("AffineFEStateMapTest.jl")
-  include("MultiFieldAffineFEStateMapTest.jl")
-  include("MultiFieldNonlinearFEStateMapTest.jl")
-  include("NonlinearFEStateMapTest.jl")
-  include("RepeatingAffineFEStateMapTest.jl")
-  include("TwoStaggeredAffineFEStateMapTest.jl")
-  include("ThreeStaggeredAffineFEStateMapTests.jl")
-  include("ThreeStaggeredNonlinearFEStateMapTest.jl")
-  include("AffineFEStateMapTest_CutFEM_Elast.jl")
-  include("Zygote_TwoStaggeredAffineFEStateMapTest.jl")
+  # include("AffineFEStateMapTest.jl")
+  # include("MultiFieldAffineFEStateMapTest.jl")
+  # include("MultiFieldNonlinearFEStateMapTest.jl")
+  # include("NonlinearFEStateMapTest.jl")
+  # include("RepeatingAffineFEStateMapTest.jl")
+  # include("TwoStaggeredAffineFEStateMapTest.jl")
+  # include("ThreeStaggeredAffineFEStateMapTests.jl")
+  # include("ThreeStaggeredNonlinearFEStateMapTest.jl")
+  # include("AffineFEStateMapTest_CutFEM_Elast.jl")
+  # include("Zygote_TwoStaggeredAffineFEStateMapTest.jl")
   include("Zygote_StaggeredMultiField.jl")
-  include("TwoStaggeredAffineFEStateMapTest_ADTypeUnstableBug.jl")
-  include("NonSymmetricThreeStaggeredNonlinearFEStateMapTest.jl")
-  # Transient
-  include("Transient_CrankNicolson_AffineFEStateMap.jl")
-  include("Transient_CrankNicolson_ConstantMats_AffineFEStateMap.jl")
-  include("Transient_CrankNicolson_NonlinearFEStateMap.jl")
+  # include("TwoStaggeredAffineFEStateMapTest_ADTypeUnstableBug.jl")
+  # include("NonSymmetricThreeStaggeredNonlinearFEStateMapTest.jl")
+  # # Transient
+  # include("Transient_CrankNicolson_AffineFEStateMap.jl")
+  # include("Transient_CrankNicolson_ConstantMats_AffineFEStateMap.jl")
+  # include("Transient_CrankNicolson_NonlinearFEStateMap.jl")
 end
 
 end # module